+++ src/main/java/site/weather/api/weather/service/WeatherService.java
@@ -9,8 +9,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Mono;
 import reactor.core.scheduler.Schedulers;
-import site.weather.api.weather.dto.response.WeatherErrorResponse;
 import site.weather.api.weather.dto.response.WeatherResponse;
+import site.weather.api.weather.error.dto.WeatherErrorResponse;
 import site.weather.api.weather.repository.WeatherSubscriptionInfoRepository;
 
 @Service
